JSON endpoints, proper menu item add and edit, fix None

+++ vagrant/practise/finalMenuItem.py
@@ -20,6 +20,8 @@ session = DBSession()
 def menuItems():
     #session = db_session()
     menuItems = session.query(MenuItem).all()
+    if menuItems is None:
+        return []
     return menuItems
     
 def menuItemNames():
@@ -28,7 +30,11 @@ def menuItemNames():
 def restaurantMenu(restaurant_id):
     #session = db_session()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).first()
-    menuItems = session.query(MenuItem).filter_by(restaurant_id = restaurant.id)
+    if restaurant is None:
+        return []
+    menuItems = session.query(MenuItem).filter_by(restaurant_id = restaurant.id) 
+    if menuItems is None:
+        return []
     return menuItems
     
 def restaurantMenuNames(restaurant_id):
@@ -42,30 +48,46 @@ def menuItemById(id):
 # def MenuItemByName(name):
     # pass #return 0
     
-def addNewMenuItem(restaurant_id, menuItem_name):
+def restaurantMenuItem(restaurant_id, menu_id):
+    items = restaurantMenu(restaurant_id)
+    item = []
+    for i in items:
+        if i.id == menu_id:
+            item.append(i)
+    return item
+
+def addNewMenuItem(restaurant_id, menuItem_name, price=None,  course=None, description=None):
     #session = db_session()
-    menuItem = MenuItem(name = menuItem_name, restaurant_id = restaurant_id)
+    menuItem = MenuItem(name = menuItem_name, restaurant_id = restaurant_id, price = price, course = course, description = description)
     session.add(menuItem)
     session.commit()
 
 def editMenuItemById(id, newItem):
     #session = db_session()
     menuItem = menuItemById(id)
+    if menuItem is not None:
+        for i in newItem:
+            print "Item", menuItem
+            print "New Item", newItem
+            print "editing item %s" % (i,)
 
-    for i in newItem:
-        print "Item", menuItem
-        print "New Item", newItem
-        print "editing item %s" % (i,)
-    if newItem['name']:
+        if newItem['name']:
             menuItem.name = newItem['name']
-    session.add(menuItem)
-    session.commit()
+        if newItem['course']:
+            menuItem.course = newItem['course']
+        if newItem['price']:
+            menuItem.price = newItem['price']
+        if newItem['description']:
+            menuItem.description = newItem['description']
+        session.add(menuItem)
+        session.commit()
     
 def deleteMenuItemById(id):
     #session = db_session()
     menuItem = menuItemById(id)
-    session.delete(menuItem)
-    session.commit()
+    if menuItem is not None:
+        session.delete(menuItem)
+        session.commit()
     
 # def deleteMenuItemByName(restaurant_id, name):
     # session = db_session()
